\name{sgPLS-internal}
\encoding{latin1}
\alias{normv}
\alias{soft.thresholding}
\alias{soft.thresholding.group}
\alias{soft.thresholding.sparse.group}
\alias{lambda.quadra}
\alias{step1.spls.sparsity}



\alias{step1.sparse.group.spls.sparsity}
\alias{step1.group.spls.sparsity}

\alias{step2.spls}

\title{Internal Functions}

\description{
Internal functions not to be used by the user.
}

