\name{sgPLS-package}
\alias{sgPLS-package}
\docType{package}
\title{
Group and Sparse Group Partial Least Square Model
}
\description{The sgPLS package provides sparse, group and sparse group
  version for PLS approaches. 
	The main functions are:  \code{\link{sPLS}} for sparse PLS, \code{\link{gPLS}} for group PLS and \code{\link{sgPLS}} for sparse group PLS.
}

\author{Benoit Liquet <b.liquet@uq.edu.au>, Pierre Lafaye de Micheaux}
\references{
Liquet Benoit, Lafaye de Micheaux Pierre, Hejblum Boris, Thiebaut Rodolphe. A group and Sparse Group Partial Least Square approach applied in Genomics context. \emph{Submitted}.

}
\keyword{ package }
\seealso{
\code{\link{sgPLS}}, \code{\link{gPLS}}
}

